requirements.txt
psycopg2-binary==2.9.9
numpy==1.24.3
Flask==2.3.3
prometheus-client==0.17.1
python-dotenv==1.0.0
